Add nodemon config for debugging/development
Added a nodemon config so that when you run the app and make changes it
automatically restarts itself. In addition, if you want to debug the
app using nodemon, you can just tell VSCode to run the debug script, and
it will automagically attach to the process, even through restarts!